+++ cmd/forklift/dev/plt/cli.go
@@ -211,6 +211,15 @@ func makeQueryReqSubcmds(category string) []*cli.Command {
 			Action:   lsRepoAction,
 		},
 		{
+			Name:     "locate-repo",
+			Aliases:  []string{"locate-repository"},
+			Category: category,
+			Usage: "Prints the absolute filesystem path of a repo available in the development " +
+				"pallet",
+			ArgsUsage: "repo_path",
+			Action:    locateRepoAction,
+		},
+		{
 			Name:      "show-repo",
 			Aliases:   []string{"show-repository"},
 			Category:  category,