+++ cmd/forklift/plt/repositories.go
@@ -48,6 +48,17 @@ func lsRepoAction(c *cli.Context) error {
 	return fcli.PrintRequiredRepos(0, plt)
 }
 
+// locate-repo
+
+func locateRepoAction(c *cli.Context) error {
+	plt, caches, err := processFullBaseArgs(c.String("workspace"), true)
+	if err != nil {
+		return err
+	}
+
+	return fcli.PrintRequiredRepoLocation(plt, caches.r, c.Args().First())
+}
+
 // show-repo
 
 func showRepoAction(c *cli.Context) error {